services/infra-version-control-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
aiokafka==0.10.0
redis==5.0.1
GitPython==3.1.41
docker==7.0.0
kubernetes==29.0.0
jinja2==3.1.3
pyyaml==6.0.1
prometheus-client==0.19.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
